+++ link_elements.c
@@ -148,7 +148,7 @@ void add_element_links(ELEMENT_LINKS *links, NAMELIST_TEXT *nltext, LINE_LIST *b
     targetList = NULL;
     /* make a list of all the unique element names that match this (possibly wildcard) target */
     do {
-      long duplic;
+      int32_t duplic;
       if (!exclude || !strlen(exclude) || !wild_match(t_context->name, exclude)) {
         targetList = SDDS_Realloc(targetList, sizeof(*targetList)*(targets+1));
         binaryInsert((void**)targetList, targets, t_context->name, strcmp, &duplic);
